Adding code files and example code of loop scheduling library serving as a submodule of RAJA.
I'm adding code files and example code showing usage of library for loop scheduling. The library is to be plugged into the forall construct in RAJA, i.e., the loop scheduling library will be called whenever a forall construct is used by RAJA. 

Note that the macro-based abstraction in the loop scheduling library, i.e., 'FORALL_BEGIN() {loop body} FORALL_END()' is analogous and simpler version of the abstractions of RAJA's FORALL construct.